+++ src/periodOf.js
@@ -1,0 +1,8 @@
+import { withScheduler } from './schedulerUtils'
+import { periodOf as periodOf_ } from './internal/periodOf'
+
+const periodOf = withScheduler(periodOf_)
+
+export {
+  periodOf,
+}